refactor: renamed hooks directory to reducers
